+++ src/libs/Detach/CHANGELOG.md
@@ -6,13 +6,13 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Added `GetCorners` method to `PyramidFrustum` (previously `Frustum`) struct.
+- Added `GetCorners` method to `ViewFrustum` (previously `Frustum`) struct.
 - Added support for `bool`, `Vector2`, `Vector3`, `Vector4`, and `Quaternion` to `InlineInterpolatedStringHandlerUtf8`.
 
 ### Changed
 
 - Changed the order of parameters in some `Geometry3D.ClosestPointIn...` methods. The `point` parameter is now always the first parameter.
-- Renamed `Frustum` to `PyramidFrustum`.
+- Renamed `Frustum` to `ViewFrustum`.
 
 ## 0.24.3